backend/requirements.txt
fastapi
uvicorn
stripe
pyjwt
